refactor user handler 하나의 클래스로 이동
